requirements.txt
attrs==21.2.0
curio==1.6
cycler==0.11.0
Cython==3.0.5
dataclasses-json==0.5.14
debugpy==1.6.6
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.6
Django==4.0
dm-tree==0.1.8
dnspython==2.4.2
docutils==0.20.1
email-validator==2.0.0.post2
entrypoints==0.4
environs==9.5.0
et-xmlfile==1.1.0
etils==1.0.0
excel-util==0.0.4
exceptiongroup==1.1.2
executing==0.9.1
fastapi==0.104.1
h11==0.14.0
h5py==3.7.0
Haskell==0.0.1
httpcore==0.16.3
httpie==3.2.2
httptools==0.6.0
httpx==0.23.3
huggingface-hub==0.16.4
idna==3.3
imageio==2.32.0
imageio-ffmpeg==0.4.8
imagesize==1.4.1
imgaug==0.4.0
imutils==0.5.4
importlib-metadata==4.12.0
importlib-resources==5.10.2
iniconfig==2.0.0
ipykernel==6.22.0
ipyparallel==8.6.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.13
jedi==0.18.1
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.3.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
keras==2.10.0
keras-nlp==0.5.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
lazy_loader==0.3
libclang==14.0.6
librosa==0.10.1
llvmlite==0.41.1
logic==0.2.3
lxml==4.7.1
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.20.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mediapipe==0.8.10.1
ml-dtypes==0.2.0
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numba==0.58.1
numexpr==2.8.5
numpy==1.22.3
oauthlib==3.2.0
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
pandas==1.4.3
pandas-util==0.0.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.1.1
prompt-toolkit==3.0.30
protobuf==3.19.6
py==1.11.0
pydantic==2.5.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.14.3
pytest==7.0.1
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
requests==2.26.0
requests-file==1.5.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
safetensors==0.3.1
scikit-image==0.22.0
scikit-learn==1.3.1
scipy==1.9.0
sklearn==0.0
sniffio==1.2.0
sqlparse==0.4.2
tenacity==8.2.3
tensorflow==2.10.1
tensorflow-datasets==4.8.1
tensorflow-estimator==2.10.0
tensorflow-examples @ git+https://github.com/tensorflow/examples.git@0b14ce1c88537b94772cc99d995d936417be6f5d
tensorflow-hub==0.13.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.12.0
tensorflow-text==2.10.0
termcolor==1.1.0
terminado==0.17.1
torch==2.1.0
torchvision==0.16.0
typing-inspect==0.9.0
typing_extensions==4.8.0
uvicorn==0.23.2
